+++ src/main/java/org/markdownwriterfx/options/OptionsDialog.java
@@ -79,11 +79,13 @@ public class OptionsDialog
 
 	private void load() {
 		generalOptionsPane.load();
+		editorOptionsPane.load();
 		markdownOptionsPane.load();
 	}
 
 	private void save() {
 		generalOptionsPane.save();
+		editorOptionsPane.save();
 		markdownOptionsPane.save();
 	}
 
@@ -92,6 +94,8 @@ public class OptionsDialog
 		tabPane = new TabPane();
 		generalTab = new Tab();
 		generalOptionsPane = new GeneralOptionsPane();
+		editorTab = new Tab();
+		editorOptionsPane = new EditorOptionsPane();
 		markdownTab = new Tab();
 		markdownOptionsPane = new MarkdownOptionsPane();
 
@@ -105,13 +109,19 @@ public class OptionsDialog
 				generalTab.setContent(generalOptionsPane);
 			}
 
+			//======== editorTab ========
+			{
+				editorTab.setText(Messages.get("OptionsDialog.editorTab.text"));
+				editorTab.setContent(editorOptionsPane);
+			}
+
 			//======== markdownTab ========
 			{
 				markdownTab.setText(Messages.get("OptionsDialog.markdownTab.text"));
 				markdownTab.setContent(markdownOptionsPane);
 			}
 
-			tabPane.getTabs().addAll(generalTab, markdownTab);
+			tabPane.getTabs().addAll(generalTab, editorTab, markdownTab);
 		}
 		// JFormDesigner - End of component initialization  //GEN-END:initComponents
 	}
@@ -120,6 +130,8 @@ public class OptionsDialog
 	private TabPane tabPane;
 	private Tab generalTab;
 	private GeneralOptionsPane generalOptionsPane;
+	private Tab editorTab;
+	private EditorOptionsPane editorOptionsPane;
 	private Tab markdownTab;
 	private MarkdownOptionsPane markdownOptionsPane;
 	// JFormDesigner - End of variables declaration  //GEN-END:variables